Implements abstract class ApiEndpoint
