+++ src/definition/api/IApi.ts
@@ -1,29 +1,5 @@
-import { IHttp, IModify, IPersistence, IRead } from '../accessors';
-import { IApiEndpointInfo } from './IApiEndpointInfo';
-import { IApiExample } from './IApiExample';
-import { IApiRequest } from './IRequest';
-import { IApiResponse } from './IResponse';
-
-export interface IApiEndpoint {
-    /**
-     * Path to complete the api URL. Example: https://{your-server-address}/api/apps/public/{your-app-id}/{path}
-     * or https://{your-server-address}/api/apps/private/{your-app-id}/{private-hash}/{path}
-     */
-    path: string;
-    examples?: {[key: string]: IApiExample};
-
-    /**
-     * Called whenever the publically accessible url for this App is called,
-     * if you handle the methods differently then split it out so your code doesn't get too big.
-     */
-    get?(request: IApiRequest, endpoint: IApiEndpointInfo, read: IRead, modify: IModify, http: IHttp, persis: IPersistence): Promise<IApiResponse>;
-    post?(request: IApiRequest, endpoint: IApiEndpointInfo, read: IRead, modify: IModify, http: IHttp, persis: IPersistence): Promise<IApiResponse>;
-    put?(request: IApiRequest, endpoint: IApiEndpointInfo, read: IRead, modify: IModify, http: IHttp, persis: IPersistence): Promise<IApiResponse>;
-    delete?(request: IApiRequest, endpoint: IApiEndpointInfo, read: IRead, modify: IModify, http: IHttp, persis: IPersistence): Promise<IApiResponse>;
-    head?(request: IApiRequest, endpoint: IApiEndpointInfo, read: IRead, modify: IModify, http: IHttp, persis: IPersistence): Promise<IApiResponse>;
-    options?(request: IApiRequest, endpoint: IApiEndpointInfo, read: IRead, modify: IModify, http: IHttp, persis: IPersistence): Promise<IApiResponse>;
-    patch?(request: IApiRequest, endpoint: IApiEndpointInfo, read: IRead, modify: IModify, http: IHttp, persis: IPersistence): Promise<IApiResponse>;
-}
+import { ApiEndpoint } from './ApiEndpoint';
+import { IApiEndpoint } from './IApiEndpoint';
 
 /**
  * Represents an api that is being provided.
@@ -40,7 +16,7 @@ export interface IApi {
     /**
      * Provide enpoints for this api registry
      */
-    endpoints: Array<IApiEndpoint>;
+    endpoints: Array<ApiEndpoint | IApiEndpoint>;
 }
 
 export enum ApiVisibility {
