+++ tests/server/managers/AppApi.spec.ts
@@ -1,7 +1,7 @@
 import { Expect, SetupFixture, Test } from 'alsatian';
 
 import { IApi } from '../../../src/definition/api';
-import { IApiEndpoint } from '../../../src/definition/api/IApi';
+import { IApiEndpoint } from '../../../src/definition/api/IApiEndpoint';
 import { AppApi } from '../../../src/server/managers/AppApi';
 import { ProxiedApp } from '../../../src/server/ProxiedApp';
